% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summed_exps.R
\name{log_summed_exps}
\alias{log_summed_exps}
\title{Compute the logarithm of the sum of exponentials}
\usage{
log_summed_exps(x)
}
\arguments{
\item{x}{A vector of type numeric}
}
\value{
The logarithm of the sum of exponentials of the input values
}
\description{
Compute the logarithm of the sum of exponentials
}
\examples{
log_summed_exps(1:2000)

}
